+++ src/trx/game/inject/enum.h
@@ -94,7 +94,8 @@ typedef enum {
     IDT_OBJ_LINK_EDITS   = 36,
     IDT_ITEM_NAME_EDITS  = 37,
     IDT_FLYBY_CAMERAS    = 38,
-    IDT_NUMBER_OF        = 39,
+    IDT_PROPERTY_EDITS   = 39,
+    IDT_NUMBER_OF        = 40,
 } INJECTION_DATA_TYPE;
 
 typedef enum {